Refactored .csv format to fix for split production before/after adjustment
Accounted for the (fairly common) case in which a plant intitially produces at partial capacity, but then is adjusted further upwards.

'Price' is no longer a meaningful field, as there is no guarantee that all of a plant's electricity will be sold at the same price. Refactored to split into base production, downwards-adjusted production, and upwards-adjusted production.