BREAKING UPDATE: Changed demand curve format
The demand curve was previously represented by the equation P = m(Q - b). This is the inverse of Q = (1/m)P + b. Students and instructors found this unnecessarily confusing (as it was — it doesn't quite line up with either the traditional demand curve or inverse demand curve models). As such, the demand curve is now represented by the equation Q = mP - b. (The inverse of this is P = 1/m(Q - b).) THIS IS A MAJOR UPDATE. The 'slope' column in schedule.csv must be updated in order to see the same outputs as prior to this update (see the update to schedule.csv in this commit for an example). Furthermore, the headers in schedule.csv have been updated to more clearly reflect the semantics of the demand equation.

Again, the relationship between Price and Quantity Demanded is as follows:
Quantity Demanded = ([schedule.csv's 'slope'] * Price) + [schedule.csv's 'net_base_demand']